optimize the payment flow change

+++ src/interfaces/IFlashAccountant.sol
@@ -39,7 +39,7 @@ interface IFlashAccountant {
     // If you want to pay in the chain's native token, simply transfer it to this contract using a call.
     // The payer must implement payCallback in which they must transfer the token to Core.
     function startPayment(address token) external returns (uint256 balance);
-    function completePayment(address token) external returns (uint128 payment);
+    function completePayment(address token) external returns (uint256 payment);
 
     // Withdraws a token amount from the accountant to the given recipient.
     // The contract must be locked, as it tracks the withdrawn amount against the current locker's delta.
